flask_backend/requirements.txt
asgiref==3.5.2
click==8.1.3
colorama==0.4.5
Django==4.1
Flask==2.2.2
Flask-Cors==3.0.10
image==1.5.33
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.2
opencv-python==4.6.0.66
Pillow==9.2.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
Werkzeug==2.2.2
